+++ src/dsp/mod.rs
@@ -11,7 +11,7 @@ pub mod oscillator;
 pub use delay_line::DelayLine;
 pub use envelope::{init_envelope, Dahdsr, DahdsrState, EnvelopeParams};
 pub use fastmath::{
-    atan2f, cosf, exp2f, expf, expm1f, fast_tan, fast_tanh, fast_tanh_f32, ftz, log2f, modpi,
+    atan2f, cosf, exp2f, fast_tan, fast_tanh, fast_tanh_f32, ftz, log2f, modpi,
     par_cosf, par_sinf, pow10, pow1half, powf, sinf,
 };
 pub use filter::{Biquad, FilterState, Svf, SvfMode, SvfState};